refactor: update API package structure to v2 for improved organization and versioning
