+++ src/main/java/org/thingai/app/api/v2/EventApi.java
@@ -1,9 +1,10 @@
-package org.thingai.app.api.v1;
+package org.thingai.app.api.v2;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api/team")
-public class TeamApi {
+@RequestMapping("/api/event")
+public class EventApi {
+
 }
