+++ src/main/java/org/thingai/app/api/v2/MatchApi.java
@@ -1,4 +1,4 @@
-package org.thingai.app.api.v1;
+package org.thingai.app.api.v2;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.http.ResponseEntity;
