+++ src/main/java/org/thingai/app/api/v2/MatchControlApi.java
@@ -1,10 +1,11 @@
-package org.thingai.app.api.v1;
+package org.thingai.app.api.v2;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api/sync")
-public class SyncApi {
+@RequestMapping("/api/scorekeeper")
+public class MatchControlApi {
+
 
 }
