+++ src/main/java/org/thingai/app/api/v2/SyncApi.java
@@ -1,10 +1,10 @@
-package org.thingai.app.api.v1;
+package org.thingai.app.api.v2;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api/rank")
-public class RankApi {
+@RequestMapping("/api/sync")
+public class SyncApi {
 
 }
