+++ src/main/java/org/thingai/app/api/v2/TeamApi.java
@@ -1,11 +1,9 @@
-package org.thingai.app.api.v1;
+package org.thingai.app.api.v2;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api/scorekeeper")
-public class MatchControlApi {
-
-
+@RequestMapping("/api/team")
+public class TeamApi {
 }
